requirements.txt
PyYAML==6.0
Jinja2==3.1.2
pydantic==1.10.2
pdfplumber==0.7.4  # For future PDF parsing